refactor: rename user.services to user.service to fix import inconsistencies
