+++ src/modules/user/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { UserController } from './user.controller';
-import { UserService } from './user.services';
+import { UserService } from './user.service';
 import { UserRepository } from './user.repository';
 
 @Module({
